requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
future==0.18.2
github==1.2.7
idna==3.4
multidict==6.0.2
python-whois==0.8.0
requests==2.28.1
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
